sendable updates
