+++ Sources/OpenWeatherKit/Internal/Extensions/Sendable+Compat.swift
@@ -1,0 +1,18 @@
+//
+//  Sendable+Compat.swift
+//  
+//
+//  Created by Jeremy Greenwood on 11/24/22.
+//
+
+import Foundation
+
+extension Date: @unchecked Sendable {}
+extension Measurement: @unchecked Sendable {}
+extension URL: @unchecked Sendable {}
+extension UnitAngle: @unchecked Sendable {}
+extension UnitLength: @unchecked Sendable {}
+extension UnitPressure: @unchecked Sendable {}
+extension UnitSpeed: @unchecked Sendable {}
+extension UnitTemperature: @unchecked Sendable {}
+extension URLSession: @unchecked Sendable {}
