+++ Sources/OpenWeatherKit/Public/Requests/WeatherMetadata.swift
@@ -21,4 +21,3 @@ public struct WeatherMetadata: Sendable {
 
 extension WeatherMetadata: Codable {}
 extension WeatherMetadata: Equatable {}
-extension Date: Sendable {}
